refactor(scheduler): 将工具类方法集中在文件util中
