+++ pkg/scheduler/scheduler.go
@@ -9,20 +9,6 @@ import (
 	"github.com/google/uuid"
 )
 
-// escape escapes backslashes and colons in a string.
-func escape(s string) string {
-	s = strings.ReplaceAll(s, "\\", "\\\\")
-	s = strings.ReplaceAll(s, ":", "\\:")
-	return s
-}
-
-// unescape unescapes backslashes and colons in a string.
-func unescape(s string) string {
-	s = strings.ReplaceAll(s, "\\\\", "\\")
-	s = strings.ReplaceAll(s, "\\:", ":")
-	return s
-}
-
 // OpScheduler is the main scheduler struct that manages jobs.
 type OpScheduler struct {
 	Name         string
